# src/http/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TMDB M-Team Hub Management API",
    "version": "1.0.0"
  },
  "security": [
    {
      "ManagementSession": []
    }
  ],
  "paths": {
...
    "/api/search": {
      "get": {
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "TMDB search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TmdbSearchResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid search request or missing TMDB credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ManagementUnauthorized"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "description": "TMDB request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
    "/api/tmdb/movie/{id}": {
      "get": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "force_refresh",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Normalized TMDB movie detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TmdbMediaDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid movie request or missing TMDB credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ManagementUnauthorized"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "description": "TMDB request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/api/tmdb/tv/{id}": {
      "get": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "force_refresh",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Normalized TMDB TV detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TmdbMediaDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid TV request or missing TMDB credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ManagementUnauthorized"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "description": "TMDB request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/api/tmdb/tv/{id}/season/{season}": {
      "get": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "season",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "force_refresh",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Normalized TMDB season detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TmdbSeasonDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid season request or missing TMDB credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/ManagementUnauthorized"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "description": "TMDB request failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ManagementSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tmdb_mteam_admin_session"
      }
    },
    "responses": {
      "ManagementUnauthorized": {
        "description": "Management authentication is required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorDto"
            }
          }
        }
      },
      "MethodNotAllowed": {
        "description": "HTTP method is not allowed for this endpoint",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorDto"
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal server error with a redacted public message",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorDto"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "A required local service or store is temporarily unavailable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorDto"
            }
          }
        }
      }
    },
    "schemas": {
...
      "TmdbMediaKind": {
        "type": "string",
        "enum": ["movie", "tv"]
      },
      "TmdbSearchItemDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "media_type",
          "id",
          "title",
          "original_title",
          "overview",
          "poster_path",
          "poster_url",
          "release_date",
          "first_air_date",
          "vote_average"
        ],
        "properties": {
          "media_type": {
            "$ref": "#/components/schemas/TmdbMediaKind"
          },
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "original_title": {
            "type": "string"
          },
          "overview": {
            "type": "string"
          },
          "poster_path": {
            "type": ["string", "null"]
          },
          "poster_url": {
            "type": ["string", "null"]
          },
          "release_date": {
            "type": ["string", "null"]
          },
          "first_air_date": {
            "type": ["string", "null"]
          },
          "vote_average": {
            "type": ["number", "null"]
          }
        }
      },
      "NamedMediaValueDto": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "name", "english_name", "code"],
        "properties": {
          "id": {
            "type": ["integer", "null"]
          },
          "name": {
            "type": "string"
          },
          "english_name": {
            "type": ["string", "null"]
          },
          "code": {
            "type": ["string", "null"]
          }
        }
      },
      "TmdbSeasonSummaryDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "season_number",
          "name",
          "episode_count",
          "air_date",
          "poster_path",
          "poster_url"
        ],
        "properties": {
          "season_number": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "episode_count": {
            "type": ["integer", "null"],
            "minimum": 0
          },
          "air_date": {
            "type": ["string", "null"]
          },
          "poster_path": {
            "type": ["string", "null"]
          },
          "poster_url": {
            "type": ["string", "null"]
          }
        }
      },
      "TmdbMediaDetailDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "media_type",
          "id",
          "title",
          "original_title",
          "overview",
          "tagline",
          "poster_path",
          "poster_url",
          "backdrop_path",
          "backdrop_url",
          "release_date",
          "first_air_date",
          "last_air_date",
          "runtime",
          "status",
          "vote_average",
          "vote_count",
          "genres",
          "production_countries",
          "spoken_languages",
          "origin_country",
          "imdb_id",
          "douban_id",
          "douban_url",
          "number_of_seasons",
          "number_of_episodes",
          "episode_run_time",
          "networks",
          "series_type",
          "seasons"
        ],
        "properties": {
          "media_type": {
            "$ref": "#/components/schemas/TmdbMediaKind"
          },
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "original_title": {
            "type": "string"
          },
          "overview": {
            "type": "string"
          },
          "tagline": {
            "type": ["string", "null"]
          },
          "poster_path": {
            "type": ["string", "null"]
          },
          "poster_url": {
            "type": ["string", "null"]
          },
          "backdrop_path": {
            "type": ["string", "null"]
          },
          "backdrop_url": {
            "type": ["string", "null"]
          },
          "release_date": {
            "type": ["string", "null"]
          },
          "first_air_date": {
            "type": ["string", "null"]
          },
          "last_air_date": {
            "type": ["string", "null"]
          },
          "runtime": {
            "type": ["integer", "null"],
            "minimum": 0
          },
          "status": {
            "type": ["string", "null"]
          },
          "vote_average": {
            "type": ["number", "null"]
          },
          "vote_count": {
            "type": ["integer", "null"],
            "minimum": 0
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NamedMediaValueDto"
            }
          },
          "production_countries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NamedMediaValueDto"
            }
          },
          "spoken_languages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NamedMediaValueDto"
            }
          },
          "origin_country": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "imdb_id": {
            "type": ["string", "null"]
          },
          "douban_id": {
            "type": ["string", "null"]
          },
          "douban_url": {
            "type": ["string", "null"]
          },
          "number_of_seasons": {
            "type": ["integer", "null"],
            "minimum": 0
          },
          "number_of_episodes": {
            "type": ["integer", "null"],
            "minimum": 0
          },
          "episode_run_time": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "networks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NamedMediaValueDto"
            }
          },
          "series_type": {
            "type": ["string", "null"]
          },
          "seasons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TmdbSeasonSummaryDto"
            }
          }
        }
      },
      "TmdbEpisodeDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "episode_number",
          "name",
          "overview",
          "air_date",
          "still_path",
          "still_url",
          "runtime",
          "vote_average"
        ],
        "properties": {
          "id": {
            "type": ["integer", "null"]
          },
          "episode_number": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "overview": {
            "type": "string"
          },
          "air_date": {
            "type": ["string", "null"]
          },
          "still_path": {
            "type": ["string", "null"]
          },
          "still_url": {
            "type": ["string", "null"]
          },
          "runtime": {
            "type": ["integer", "null"],
            "minimum": 0
          },
          "vote_average": {
            "type": ["number", "null"]
          }
        }
      },
      "TmdbSeasonDetailDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "season_number",
          "name",
          "overview",
          "air_date",
          "poster_path",
          "poster_url",
          "episodes"
        ],
        "properties": {
          "id": {
            "type": ["integer", "null"]
          },
          "season_number": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "overview": {
            "type": "string"
          },
          "air_date": {
            "type": ["string", "null"]
          },
          "poster_path": {
            "type": ["string", "null"]
          },
          "poster_url": {
            "type": ["string", "null"]
          },
          "episodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TmdbEpisodeDto"
            }
          }
        }
      },
...
      "TmdbSearchResponseDto": {
        "type": "object",
        "additionalProperties": false,
        "required": ["movies", "tv"],
        "properties": {
          "movies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TmdbSearchItemDto"
            }
          },
          "tv": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TmdbSearchItemDto"
            }
          }
        }
      },
...
      "ApiErrorDto": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "message"],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}